Refactoring to display The Movie Database
