+++ .idea/libraries/com_android_support_leanback_v17_26_1_0.xml
@@ -1,0 +1,12 @@
+<component name="libraryTable">
+  <library name="com.android.support:leanback-v17-26.1.0">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/leanback-v17-26.1.0.aar/61006e598c87e15c808868244c1c0f44/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/leanback-v17-26.1.0.aar/61006e598c87e15c808868244c1c0f44/res" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.android.support/leanback-v17/26.1.0/3fca2022896ef9a48932d4997fbbec59a9c31d0a/leanback-v17-26.1.0-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
