+++ app/src/main/res/layout/activity_details.xml
@@ -2,8 +2,8 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/details_fragment"
-    android:name="com.amohnacs.bitcointv.details.ui.VideoDetailsFragment"
+    android:name="com.amohnacs.moviemenu.details.ui.VideoDetailsFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.amohnacs.bitcointv.details.ui.DetailsActivity"
+    tools:context="com.amohnacs.moviemenu.details.ui.DetailsActivity"
     tools:deviceIds="tv" />
