+++ app/src/main/res/layout/activity_main.xml
@@ -2,9 +2,9 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main_browse_fragment"
-    android:name="com.amohnacs.bitcointv.main.ui.MainFragment"
+    android:name="com.amohnacs.moviemenu.main.ui.MainFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.amohnacs.bitcointv.main.ui.MainActivity"
+    tools:context="com.amohnacs.moviemenu.main.ui.MainActivity"
     tools:deviceIds="tv"
     tools:ignore="MergeRootFrame" />
