+++ app/src/main/res/values/colors.xml
@@ -1,7 +1,14 @@
 <resources>
+    <color name="primary">#0096a6</color>
+    <color name="primary_light">#B3E5FC</color>
+    <color name="primary_dark">#0288D1</color>
+    <color name="accent">#FFC107</color>
+    <color name="text_primary">#212121</color>
+    <color name="text_secondary">#757575</color>
+
     <color name="background_gradient_start">#000000</color>
     <color name="background_gradient_end">#DDDDDD</color>
-    <color name="fastlane_background">#0096a6</color>
+    <color name="fastlane_background">@color/primary</color>
     <color name="search_opaque">#ffaa3f</color>
     <color name="selected_background">#ffaa3f</color>
     <color name="default_background">#3d3d3d</color>
